requirements.txt
fusion-solar-py==0.1.1
paho-mqtt>=2.0.0
